add redirects back up to root
The problem is that if there's both a "news.html" file and a "news/"
dir, GitHub would rather try to serve the dir and die with a "can't
generate index" error than serve up the html file.  These files bounce
the user back to the correct HTML file without error.